beautirag-app/src/backend/requirements.txt
# Notebook test
ipywidgets
ipykernel
jupyter

# Core Frameworks
fastapi
uvicorn[standard]
python-dotenv

# LangChain & Vector DB
langchain
langchain-community 
faiss-cpu 
langchain_openai
langchain_anthropic
langchain_deepseek

# Document Loaders
pypdf
python-docx
docx2txt
libmagic
unstructured[local-inference]
unstructured # handles various formats


# Image Processing (OCR)
Pillow
pytesseract

# Audio Processing (Transcription)
# Whisper requires ffmpeg to be installed on the system
openai-whisper

# Embeddings
sentence-transformers

# API Specific
python-multipart # For FastAPI file uploads